+++ demo/index.js
@@ -50,8 +50,8 @@ if (window.innerWidth < 768) {
 }
 
 window.vditor = new Vditor('vditor', {
-  // _lutePath: `http://192.168.0.107:9090/lute.min.js?${new Date().getTime()}`,
-  _lutePath: 'src/js/lute/lute.min.js',
+  _lutePath: `http://192.168.0.107:9090/lute.min.js?${new Date().getTime()}`,
+  // _lutePath: 'src/js/lute/lute.min.js',
   toolbar,
   mode: 'wysiwyg',
   height: window.innerHeight + 100,
@@ -65,6 +65,9 @@ window.vditor = new Vditor('vditor', {
       mark: true,
       footnotes: true,
     },
+    math: {
+      engine: 'MathJax',
+    }
   },
   toolbarConfig: {
     pin: true,