+++ src/ts/export/index.ts
@@ -67,7 +67,7 @@ export const exportHTML = (vditor: IVditor) => {
         cdn: '${vditor.options.cdn}',
         math: ${JSON.stringify(vditor.options.preview.math)},
     });
-    Vditor.mermaidRender(previewElement, '${vditor.options.cdn}');
+    Vditor.mermaidRender(previewElement, '${vditor.options.cdn}', '${vditor.options.theme}');
     Vditor.flowchartRender(previewElement, '${vditor.options.cdn}');
     Vditor.graphvizRender(previewElement, '${vditor.options.cdn}');
     Vditor.chartRender(previewElement, '${vditor.options.cdn}');