+++ src/ts/util/processCode.ts
@@ -67,7 +67,7 @@ export const processCodeRender = (previewPanel: HTMLElement, vditor: IVditor) =>
     if (language === "abc") {
         abcRender(previewPanel, vditor.options.cdn);
     } else if (language === "mermaid") {
-        mermaidRender(previewPanel, vditor.options.cdn);
+        mermaidRender(previewPanel, vditor.options.cdn, vditor.options.theme);
     } else if (language === "flowchart") {
         flowchartRender(previewPanel, vditor.options.cdn);
     } else if (language === "echarts") {