+++ CHANGELOG.md
@@ -83,6 +83,8 @@
 
 ### v3.6.4 / 2020-11-xx
 
+* [819](https://github.com/Vanessa219/vditor/issues/819) 粘贴后滚动到粘贴内容末尾 bug `修复缺陷`
+
 ### v3.6.3 / 2020-11-19
 
 * [815](https://github.com/Vanessa219/vditor/issues/815) wysiwyg sv 表格切换会多内容 `修复缺陷`